+++ cli/__init__.py
@@ -6,7 +6,8 @@ for external service integration.
 """
 
 import sys
+
 from .main import cli
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     cli(sys.argv[1:], obj={})